- db changes and cartvm changes - auto db context changes

+++ uhrenWelt-DB-QRY.sql
@@ -25,7 +25,7 @@ CREATE TABLE [dbo].[OrderLine](
     [Id] [int] PRIMARY KEY IDENTITY(1,1) NOT NULL,
     [OrderId] [int] NOT NULL,
     [ProductId] [int] NOT NULL,
-    [Amount] [tinyint] NOT NULL,
+    [Amount] [int] NOT NULL,
     [NetUnitPrice] [decimal](12, 2) NOT NULL,
     [TaxRate] [decimal](9, 2) NOT NULL,
 )
